+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 #### _***All Apps and Devices***_
 - [NEW] Initial Release
+- [10/23/2025] 1.0.2 Fix a bug due to the recent change in Hubitat FW 2.4.3.137.
+  Reference: https://community.hubitat.com/t/hubitatalarm-dsc-integration-broken-with-2-4-3-137/158070
+
 - [02/08/2022] 1.0.1 Fix chimeToggle status parser.
 
 #### _***Alarm (Node.JS Server)***_